Added Counting Math Problem
Added a class called CountingProblems that extend MathProblems and fulfills the counting requirement (counting up to 120).

it is called just like MathProblems but using CountingProblems and I've left an example of what to change in the BasicAdditionQuizController.

I've also added a counting quiz class and fxml that is not tied to a home page button yet but does work if a button is modified to call it.